backend/requirements.txt
fastapi
uvicorn[standard]
mediapipe
pillow
numpy
python-multipart
